+++ BlueSTSDK/BlueSTSDK/Features/BlueSTSDKFeatureBattery.m
@@ -164,7 +164,7 @@ static NSArray *sFieldDesc;
 *  @throw exception if there are no 7 bytes available in the rawdata array
  *  @return battery information + number of read bytes (7)
  */
--(BlueSTSDKExtractResult*) extractData:(uint32_t)timestamp data:(NSData*)rawData dataOffset:(uint32_t)offset{
+-(BlueSTSDKExtractResult*) extractData:(uint64_t)timestamp data:(NSData*)rawData dataOffset:(uint32_t)offset{
     
     
     if(rawData.length-offset < 7){
